backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
redis>=5.0.3

# --- User Specified Composio Stack ---
composio-client==1.5.0
composio_core==0.7.15
composio_openai==0.7.20
openai>=1.0.0 # Required by composio_openai